android java package rename: com.github.realxtend -> org.realxtend.tundra
just happened to see this while reading, haven't tested building for android (yet).
figured that github and google (if we submit to Play) don't like it if we pose as made by github..

googled this a bit and apparently other github projects have it like this too so i guess it's fine. anyhow as we have our own domain too i think this is more correct.

did a search to find all occurrences and tried to move the dir accordingly.
but can't be 100% that's correct as couldn't build for android yet.